This package provides a set of functions and classes using the S version 4 class system.

FLEDA is a part of the FLR, for exploratory analysis of the data available for stock assessment. It includes a combination of simple calculations and graphical representations aiming at data screening (checking for missing data, unusual values, patterns, etc), inspection of data consistency (within and between data series) and extracting signals from the basic data. Diagnostics include those recommended during the 2004 Methods Working Group meeting (ICES, 2004).