Class "bmass" for mature and immature biomass

This class encapsulates results of a mature (mb) and immature (ib) biomass method.

Creating objects from the class

Objects of this classe should be created with a call to the creator function bmass(object, ...), where object must be of class "FLStock" and ... can be used to pass trim arguments to the creator funtion.

Slots

.Data

Object of class "FLQuants" with a mb component and a ib component.

names

Names of the list elements. character.

desc

Description of the object. character.

lock

Lock mechanism, if turned on the length of the list can not be modified by adding or removing elements. logical.

Code: .Data desc lock names

Extends

Class "FLQuants", directly.

Methods

Inherits from "FLQuants".

Details

Considering \(C_{ay}\), the catch in numbers, \(W_{ay}\), the mean weight and \(M_{ay}\), the proportion of mature fish, at age \(a=1,\dots,A\) and year \(y=1,\dots,Y\) obtained e.g. from a "FLStock" object, the computations of bmass are:

mb: $$MB_{y}=\sum_{a}{C_{ay}*W_{ay}*M_{ay}}$$

ib: $$IB_{y}=\sum_{a}{C_{ay}*W_{ay}*(1-M_{ay})}$$

See also

FLQuant-class, trim, logr-class

Examples

data(ple4)
# compute mature and immature biomass
ple4.bmass <- bmass(ple4)

#> Error in validObject(.Object): invalid class “FLQuants” object: Components must be FLQuant

# tune plot
ttl <- list(label="Trends in biomass for mature and immature plaice in IV", cex=1)
yttl <- list(label="relative biomass", cex=0.8)
xttl <- list(cex=0.8)
ax <- list(cex=0.8)
akey <- simpleKey(text=c("mature", "immature"), points=FALSE, lines=TRUE)
# plot
xyplot(data~year, data=ple4.bmass, type="l", main=ttl,
       key=akey, ylab=yttl, xlab=xttl, scales=ax)

#> Error in xyplot(data ~ year, data = ple4.bmass, type = "l", main = ttl,     key = akey, ylab = yttl, xlab = xttl, scales = ax): object 'ple4.bmass' not found