Bubbles plots for "spay" and "FLCohort" objects

Implements bubbles plots for "spay" and "FLCohort" classes.

Generic Function

See bubbles-methods.

Methods

x="formula", data="spay"

This method differs from the usual bubbles methods by using different colors for positive and negative values.

x="formula", data="FLCohort"

This method implements the usual bubbles plot to a FLCohort object.

See also

spay-class, bubbles-methods, lattice, FLCohort-class

Examples

data(ple4sex)
# compute standardized catch proportion at age
ple4sex.spay <- spay(ple4sex@catch.n)

#> Error in UseMethod("units<-"): no applicable method for 'units<-' applied to an object of class "c('FLQuant', 'FLArray', 'array', 'structure', 'vector')"

# fine tune 
ttl <- list(label="Standardized catch proportion at age for Plaice in IV", cex=1)
yttl <- list(label="age", cex=0.8)
xttl <- list(cex=0.8)
ax <- list(cex=0.7)
# plot
bubbles(age~year|unit, ple4sex.spay,  main=ttl,
        ylab=yttl, xlab=xttl, scales=ax)

#> Error in bubbles(age ~ year | unit, ple4sex.spay, main = ttl, ylab = yttl,     xlab = xttl, scales = ax): object 'ple4sex.spay' not found