Correlation for class "FLQuant"

Implements the cor method for class "FLQuant".

Methods

x="FLQuant"

This method computes correlation between ages (or whatever quant is).

x="FLQuant",y="FLQuant"

This method computes cross correlation of 2 FLQuant between ages (or whatever quant is).

Details

The cor method for one "FLQuant" can be computed between the complete vector of observed values-at-quant or can be computed between the cohort adjusted vector. The choice is made by the argument use. The first method (defined by use="all.obs") uses all observations of each two quant rows to compute cor. The second (use="complete.obs", the default) uses only the observations of each two quant rows that have a matching cohort.

This kind of analysis is trying to give a measure of the data consistency along cohorts and if the observations that have no matching cohort are left in, the correlation coeficient is contaminated by them and is no longer a good indicator.

See also

FLQuant-class, cor

Examples

data(ple4)
c1 <- cor(catch.n(ple4), use="complete.obs", method="spearman")
round(c1,2)

#> , , unit = unique, season = all, area = unique, iter = 1
#> 
#>     age
#> age      1     2     3     4    5    6    7    8    9 10
#>   1   1.00    NA    NA    NA   NA   NA   NA   NA   NA NA
#>   2   0.73  1.00    NA    NA   NA   NA   NA   NA   NA NA
#>   3   0.65  0.93  1.00    NA   NA   NA   NA   NA   NA NA
#>   4   0.61  0.89  0.85  1.00   NA   NA   NA   NA   NA NA
#>   5   0.38  0.57  0.58  0.73 1.00   NA   NA   NA   NA NA
#>   6   0.13  0.39  0.42  0.58 0.88 1.00   NA   NA   NA NA
#>   7   0.06  0.31  0.37  0.48 0.83 0.95 1.00   NA   NA NA
#>   8  -0.12  0.03  0.10  0.23 0.64 0.83 0.87 1.00   NA NA
#>   9  -0.23 -0.11 -0.06  0.09 0.54 0.69 0.76 0.90 1.00 NA
#>   10 -0.42 -0.43 -0.35 -0.22 0.26 0.48 0.54 0.76 0.84  1
#> 

c2 <- cor(catch.n(ple4), use="all.obs", method="spearman")
round(c2,2)

#> , , unit = unique, season = all, area = unique, iter = 1
#> 
#>     age
#> age      1     2     3     4    5    6    7    8    9 10
#>   1   1.00    NA    NA    NA   NA   NA   NA   NA   NA NA
#>   2   0.47  1.00    NA    NA   NA   NA   NA   NA   NA NA
#>   3   0.41  0.55  1.00    NA   NA   NA   NA   NA   NA NA
#>   4   0.14  0.27  0.42  1.00   NA   NA   NA   NA   NA NA
#>   5   0.08  0.04  0.13  0.30 1.00   NA   NA   NA   NA NA
#>   6   0.00 -0.07 -0.03  0.02 0.45 1.00   NA   NA   NA NA
#>   7  -0.27 -0.14 -0.11 -0.10 0.24 0.48 1.00   NA   NA NA
#>   8  -0.40 -0.55 -0.22 -0.17 0.14 0.24 0.52 1.00   NA NA
#>   9  -0.38 -0.58 -0.33 -0.18 0.18 0.33 0.46 0.77 1.00 NA
#>   10 -0.22 -0.45 -0.31 -0.24 0.21 0.37 0.45 0.72 0.85  1
#> 

all.equal(c1,c2)

#> [1] "Mean relative difference: 0.6557645"