Class "pay" for proportion at age per year

This class encapsulates results of a pay method.

Creating Objects from the Class

Objects of this classe should be created with a call to the contructor function pay(object, ...), where object must be an object of class "FLQuant" and ... can be used to pass trim arguments to the creator funtion.

Slots

.Data

Object of class "FLQuant".

units

Units of measurement, "character".

Extends

Class "FLQuant", directly.

Methods

Inherits from "FLQuant".

Details

Considering \(C_{ay}\), the catch in numbers at age \(a=1,\dots,A\) per year \(y=1,\dots,Y\), obtained e.g. from the catch.n slot of a "FLStock" object, the computation of pay \(P_{ay}\) is:

$$P_{ay}=\frac{C_{ay}}{\sum_a{C_{ay}}}$$

See also

FLQuant-class, trim

Examples

data(ple4sex)
# compute catch proportions at age
ple4sex.pay <- pay(ple4sex@catch.n)

#> Error in UseMethod("units<-"): no applicable method for 'units<-' applied to an object of class "c('FLQuant', 'FLArray', 'array', 'structure', 'vector')"

# fine tune 
ttl <- list(label="Catch proportion at age for Plaice in IV", cex=1)
yttl <- list(label="age", cex=0.8)
xttl <- list(cex=0.8)
ax <- list(cex=0.7)
# plot
bubbles(age~year|unit, ple4sex.pay,  main=ttl,
        ylab=yttl, xlab=xttl, scales=ax)

#> Error in bubbles(age ~ year | unit, ple4sex.pay, main = ttl, ylab = yttl,     xlab = xttl, scales = ax): object 'ple4sex.pay' not found