Class "z" for total mortality

This class encapsulates results of a z method.

Creating Objects from the Class

Objects of this classe should be created with a call to the contructor function z(object, agerng, ...), where object must be an object of class "FLQuant", agerng is the age range of fully expoited ages and ... can be used to pass trim arguments to the creator funtion.

Slots

zy

Object of class "FLQuant".

za

Object of class "FLQuant".

zc

Object of class "FLQuant".

agerng

Object of class "numeric".

Methods

show

signature(object = "z"): Display object briefly.

summary

signature(object = "z"): Displays the average total mortality computed by age and cohort.

t.test

signature(object = "z"): Tests the similarities between the estimates by year and by cohort.

xyplot

signature(object = "z"): Lattice xyplot method.

Details

Considering \(C_{ay}\), the catch in numbers at age \(a=1,\dots,A\) per year \(y=1,\dots,Y\), obtained e.g. from the catch.n slot of a "FLStock" object, the computation of z is performed by year \(Z_{y}\), by age \(Z_{a}\) and by cohorts \(Z_{c}\), for a defined age range (agerng):

Consider \(a:a \in agerng\), \(n=\#a\) and

$$L_{ay}=log{\frac{C_{ay}}{C_{a+1,y+1}}}$$

than

$$Z_{y}=\frac{\sum_{a}{L_{ay}}}{n}$$

$$Z_{a}=\frac{\sum_{y}{L_{ay}}}{Y}$$

$$Z_{c}=\frac{\sum_{c}{L_{ay}}}{n}$$

If agerng is not provided, the method will try to guess this argument by finding the ages that have less than 10% negative logs along the time series. This procedure may have unexpected side effects !

See also

FLQuant-class, trim, logr-class

Examples

data(ple4)
# compute Z
ple4z <- z(catch.n(ple4), agerng=3:6)

#> Error in dimnames(za)$year <- "all": length of 'dimnames' [6] must match that of 'dims' [5]

# tune plot
ttl <- list("Total mortality (Z) for Plaice in IV", cex=1)
xttl <- list(cex=0.8)
yttl <- list("Mean Z", cex=0.8)
# plot z by age along years
xyplot(data~year, data=ple4z@zy, type="l", main=ttl, ylab=yttl, xlab=xttl)

#> Error in xyplot(data ~ year, data = ple4z@zy, type = "l", main = ttl,     ylab = yttl, xlab = xttl): object 'ple4z' not found

# plot z by age along ages
xyplot(data~age, data=ple4z@za, type="l", main=ttl, ylab=yttl, xlab=xttl)

#> Error in xyplot(data ~ age, data = ple4z@za, type = "l", main = ttl, ylab = yttl,     xlab = xttl): object 'ple4z' not found

# plot z by cohort along years
xyplot(data~cohort, data=ple4z@zc, type="l", main=ttl, ylab=yttl, xlab=xttl)

#> Error in xyplot(data ~ cohort, data = ple4z@zc, type = "l", main = ttl,     ylab = yttl, xlab = xttl): object 'ple4z' not found