Class "bmass" for mature and immature biomass

This class encapsulates results of a mature (mb) and immature (ib) biomass method.

Creating objects from the class

Objects of this classe should be created with a call to the creator function bmass(object, ...), where object must be of class "FLStock" and ... can be used to pass trim arguments to the creator funtion.

Slots

.Data

Object of class "FLQuants" with a mb component and a ib component.

names

Names of the list elements. character.

desc

Description of the object. character.

lock

Lock mechanism, if turned on the length of the list can not be modified by adding or removing elements. logical.

Code: .Data desc lock names
Extends

Class "FLQuants", directly.

Methods

Inherits from "FLQuants".

Details

Considering \(C_{ay}\), the catch in numbers, \(W_{ay}\), the mean weight and \(M_{ay}\), the proportion of mature fish, at age \(a=1,\dots,A\) and year \(y=1,\dots,Y\) obtained e.g. from a "FLStock" object, the computations of bmass are:

mb: $$MB_{y}=\sum_{a}{C_{ay}*W_{ay}*M_{ay}}$$

ib: $$IB_{y}=\sum_{a}{C_{ay}*W_{ay}*(1-M_{ay})}$$

See also

Examples

data(ple4)
# compute mature and immature biomass
ple4.bmass <- bmass(ple4)
# tune plot
ttl <- list(label="Trends in biomass for mature and immature plaice in IV", cex=1)
yttl <- list(label="relative biomass", cex=0.8)
xttl <- list(cex=0.8)
ax <- list(cex=0.8)
akey <- simpleKey(text=c("mature", "immature"), points=FALSE, lines=TRUE)
# plot
xyplot(data~year, data=ple4.bmass, type="l", main=ttl,
       key=akey, ylab=yttl, xlab=xttl, scales=ax)

#> Error in xyplot.formula(type = "l", main = ttl, key = akey, ylab = yttl,     xlab = xttl, scales = ax, data = structure(list(age = c("all",     "all", "all", "all", "all", "all", "all", "all", "all", "all",     "all", "all", "all", "all", "all", "all", "all", "all", "all",     "all", "all", "all", "all", "all", "all", "all", "all", "all",     "all", "all", "all", "all", "all", "all", "all", "all", "all",     "all", "all", "all", "all", "all", "all", "all", "all", "all",     "all", "all", "all", "all", "all", "all", "all", "all", "all",     "all", "all", "all", "all", "all", "all", "all", "all", "all",     "all", "all", "all", "all", "all", "all", "all", "all", "all",     "all", "all", "all", "all", "all", "all", "all", "all", "all",     "all", "all", "all", "all", "all", "all", "all", "all", "all",     "all", "all", "all", "all", "all", "all", "all", "all", "all",     "all", "all", "all", "all", "all", "all", "all", "all", "all",     "all", "all", "all", "all", "all", "all", "all", "all", "all",     "all", "all", "all", "all"), year = c(1957, 1958, 1959, 1960,     1961, 1962, 1963, 1964, 1965, 1966, 1967, 1968, 1969, 1970,     1971, 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1979, 1980,     1981, 1982, 1983, 1984, 1985, 1986, 1987, 1988, 1989, 1990,     1991, 1992, 1993, 1994, 1995, 1996, 1997, 1998, 1999, 2000,     2001, 2002, 2003, 2004, 2005, 2006, 2007, 2008, 2009, 2010,     2011, 2012, 2013, 2014, 2015, 2016, 2017, 1957, 1958, 1959,     1960, 1961, 1962, 1963, 1964, 1965, 1966, 1967, 1968, 1969,     1970, 1971, 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1979,     1980, 1981, 1982, 1983, 1984, 1985, 1986, 1987, 1988, 1989,     1990, 1991, 1992, 1993, 1994, 1995, 1996, 1997, 1998, 1999,     2000, 2001, 2002, 2003, 2004, 2005, 2006, 2007, 2008, 2009,     2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017), unit = structure(c(1L,     1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,     1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,     1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,     1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,     1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,     1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,     1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,     1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,     1L), .Label = "unique", class = "factor"), season = structure(c(1L,     1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,     1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,     1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,     1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,     1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,     1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,     1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,     1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,     1L), .Label = "all", class = "factor"), area = structure(c(1L,     1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,     1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,     1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,     1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,     1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,     1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,     1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,     1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,     1L), .Label = "unique", class = "factor"), iter = structure(c(1L,     1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,     1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,     1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,     1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,     1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,     1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,     1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,     1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,     1L), .Label = "1", class = "factor"), data = c(0.574483950975233,     0.661706034886262, 0.734689437194579, 0.801902958193817,     0.825872326059766, 0.901258595773367, 1.01914246427667, 1.05232375722258,     1.03239102991567, 1.04164323256394, 1.17569400814659, 1.12307835027356,     1.05897553845129, 0.976352767090256, 0.974840895816129, 1.04352955256172,     1.03178450371452, 0.998266881807762, 0.954859365239549, 1.0612382068196,     0.98642762348775, 1.06657301485368, 0.995599332851445, 1.09121944250894,     1.0771244871909, 1.05495218833064, 1.17979850537516, 1.23627305166157,     1.4360841090824, 1.44106936681189, 1.7003874318464, 1.73500728397897,     1.8902642458808, 1.62149789771495, 1.42149800113693, 1.23488318460484,     1.14948391532124, 0.982130153761874, 0.835609756356945, 0.798008598280227,     0.834482520467431, 0.957142400676134, 0.91418501556267, 0.98314864814693,     0.798981429634099, 0.819233774345585, 0.859707408324939,     0.722948188666731, 0.67340344109669, 0.615209008668641, 0.597645728023386,     0.636044906328993, 0.6487766199133, 0.711687957060879, 0.76422092258407,     0.852868030411828, 0.923569070045933, 0.952575612733327,     0.952888803445956, 0.905017746558135, 0.898337319283907,     0.263746851779974, 0.27085796299425, 0.44854156597076, 0.560377717963613,     0.533283318584835, 0.488869545600958, 0.508200627598153,     0.57263200240216, 0.710803075675854, 0.929701073118789, 0.407975461731559,     0.440244680588067, 0.524916411330625, 0.525263329487483,     0.632227220949173, 0.636799311437273, 0.714771862104291,     0.946561770755312, 1.22453708626664, 1.18344401016263, 1.15281147263849,     1.22070994373579, 1.28133863261208, 1.30277223856706, 1.33490883807076,     1.57008755700377, 1.69816158047313, 1.90146633197763, 1.8042349670454,     2.5105782979472, 2.49274349615197, 2.55432770706432, 1.6327916697431,     1.40269031680387, 1.19590302932051, 1.10937953392267, 1.03729945883728,     0.837507809139183, 0.800785772634956, 0.877902397775035,     1.24134735035261, 1.34942945786234, 1.4478383779287, 0.717428580571602,     0.794579149972762, 1.09326714520367, 1.21700733803059, 1.25892323947203,     0.81684238378376, 0.907581257238631, 0.794425735037761, 0.887855173968375,     0.906435668209122, 0.813203305754366, 0.697041022786626,     0.739123927939278, 0.7612971741965, 0.6569213362247, 0.602165220857466,     0.589301068592793, 0.465830148049714), qname = structure(c(1L,     1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,     1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,     1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,     1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,     2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L,     2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L,     2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L,     2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L,     2L), .Label = c("mb", "ib"), class = "factor")), row.names = c(NA,     -122L), units = c(mb = "t", ib = "t"), class = "data.frame"),     x = data ~ year): object 'ttl' not found