Correlation for class "FLQuant"

Implements the cor method for class "FLQuant".

Methods

x="FLQuant"

This method computes correlation between ages (or whatever quant is).

x="FLQuant",y="FLQuant"

This method computes cross correlation of 2 FLQuant between ages (or whatever quant is).

Details

The cor method for one "FLQuant" can be computed between the complete vector of observed values-at-quant or can be computed between the cohort adjusted vector. The choice is made by the argument use. The first method (defined by use="all.obs") uses all observations of each two quant rows to compute cor. The second (use="complete.obs", the default) uses only the observations of each two quant rows that have a matching cohort.

This kind of analysis is trying to give a measure of the data consistency along cohorts and if the observations that have no matching cohort are left in, the correlation coeficient is contaminated by them and is no longer a good indicator.

See also

Examples

data(ple4)
c1 <- cor(catch.n(ple4), use="complete.obs", method="spearman")
round(c1,2)

#> , , unit = unique, season = all, area = unique, iter = 1
#> 
#>     age
#> age      1     2     3     4    5    6    7    8    9 10
#>   1   1.00    NA    NA    NA   NA   NA   NA   NA   NA NA
#>   2   0.89  1.00    NA    NA   NA   NA   NA   NA   NA NA
#>   3   0.77  0.94  1.00    NA   NA   NA   NA   NA   NA NA
#>   4   0.73  0.86  0.89  1.00   NA   NA   NA   NA   NA NA
#>   5   0.59  0.69  0.71  0.93 1.00   NA   NA   NA   NA NA
#>   6   0.44  0.53  0.56  0.81 0.96 1.00   NA   NA   NA NA
#>   7   0.30  0.38  0.41  0.64 0.85 0.95 1.00   NA   NA NA
#>   8   0.11  0.13  0.18  0.44 0.69 0.83 0.94 1.00   NA NA
#>   9  -0.02 -0.07 -0.03  0.24 0.52 0.68 0.80 0.94 1.00 NA
#>   10 -0.33 -0.41 -0.40 -0.18 0.12 0.33 0.52 0.73 0.85  1
#> 

c2 <- cor(catch.n(ple4), use="all.obs", method="spearman")
round(c2,2)

#> , , unit = unique, season = all, area = unique, iter = 1
#> 
#>     age
#> age      1     2     3     4    5    6    7    8    9 10
#>   1   1.00    NA    NA    NA   NA   NA   NA   NA   NA NA
#>   2   0.51  1.00    NA    NA   NA   NA   NA   NA   NA NA
#>   3   0.36  0.51  1.00    NA   NA   NA   NA   NA   NA NA
#>   4   0.26  0.23  0.36  1.00   NA   NA   NA   NA   NA NA
#>   5   0.08  0.08  0.14  0.45 1.00   NA   NA   NA   NA NA
#>   6   0.00 -0.01  0.08  0.19 0.51 1.00   NA   NA   NA NA
#>   7  -0.28 -0.16 -0.13  0.01 0.20 0.47 1.00   NA   NA NA
#>   8  -0.38 -0.44 -0.30 -0.14 0.03 0.21 0.49 1.00   NA NA
#>   9  -0.35 -0.41 -0.38 -0.20 0.03 0.17 0.32 0.65 1.00 NA
#>   10 -0.22 -0.40 -0.40 -0.24 0.03 0.22 0.40 0.65 0.81  1
#> 

all.equal(c1,c2)

#> [1] "Mean relative difference: 0.7374044"